SCRUM-25-backend-associate-a-jira-project-with-an-internal-project SCRUM-24 backend create a new project SCRUM-23 backend retrieve project for id SCRUM-22 backend retrieve all projects
