+++ controllers/v1/routes/frontend/meetings/index.js
@@ -1,0 +1,9 @@
+'use strict';
+
+const express = require('express');
+const router = express.Router();
+const meetingsRoutes = require('./meetings');
+
+meetingsRoutes(router);
+
+module.exports = router;
