--- controllers/v1/routes/frontend/project/index.js
@@ -1,15 +1,0 @@
-'use strict';
-
-const express = require('express');
-const router = express.Router();
-
-// Import meetings routes
-const meetings = require('./meetings');
-
-// Use the meetings routes
-router.use('/meetings', meetings);
-
-// Export the routes in the expected structure
-module.exports = {
-    meetings: meetings
-};
